expr: Combine multiple ipv4 with wildcard mask.
This patch merges ipv4 addresses with wildcard masks, and replaces this
ipv4 addresses with the combined ip/mask. This will greatly reduce the
entries in the ovs security group flow table, especially when the host
size is large.

Analysis in the simplest scenario, a network 1.1.1.0/24 network,create 253
ports(1.1.1.2-1.1.1.254).
Only focus on the number of ip addresses, the original 253 addresses will
be combined into 13 addresses.
1.1.1.2/31
1.1.1.4/30
1.1.1.8/29
1.1.1.16/28
1.1.1.32/27
1.1.1.64/26
1.1.1.128/26
1.1.1.192/27
1.1.1.224/28
1.1.1.240/29
1.1.1.248/30
1.1.1.252/31
1.1.1.254

Some scenes are similar to the following:
    1.1.1.2, 1.1.1.6
After the combine:
    1.1.1.2/255.255.255.251
You can use ovn-match-ip utility to match ip.
such as:
    ovs-ofctl dump-flows br-int | ovn-match-ip 1.1.1.6
    1.1.1.2/255.255.255.251 will show.

Simple description of the algorithm.
There are two situations
1. Combine once
such as:
    1.1.1.0 1.1.1.1 1.0.1.0 1.0.1.1
    Combined into: 1.1.1.0/31, 1.0.1.0/31
2. Combine multiple times
    1.1.1.0 1.1.1.1 1.0.1.0 1.0.1.1
    Combined into: 1.0.1.0/255.254.255.254

Considering the actual scene and simplicity, the first case is used to
combine once.

...00...
...01...
...10...
...11...
"..." means the same value omitted.
Obviously, the above value can be expressed as ...00.../11100111. This
continuous interval that can be represented by one or several wildcard
masks is called a segment.
Only if all 2<<n values from the minimum value 00...(n) to 11...(n)
exist, can they be combined into 00...(n)/00...( n)

First sort all the values by size. Iterate through each value.
1. Find a new segment, where two values differ only by 1 bit, such as
...0... and ...1...
    diff = ip_next ^ ip
    if (diff & (diff-1)) == 0
        new_segment = true
The first non-zero place in the high direction of ip is the end of the
segment(segment_end).
For example...100... and...101..., the segment_end is ...111...

2. Count the number of consecutive and less than continuous_size in the
segment.
    diff = ip_next - ip
    if (diff & (diff-1)) == 0 && ip_next <= segment_end
        continuous_size++

3. Combine different ip intervals in the segment according to
continuous_size.
In continuous_size, from the highest bit of 1 to the lowest bit of 1, in
the order of segment start, each bit that is 1 is a different ip interval
that can be combined with a wildcard mask.
For example, 000, 001, 010:
    continuous_size: 3 (binary 11), segment_start: 000
    mask: ~(1 << 1 - 1) = 110; ~(1 << 0 - 1) = 111;
    Combined to: 000/110, 010/111

4. The ip that cannot be recorded in a segment will not be combined.

Signed-off-by: Gongming Chen <gmingchen@tencent.com>